// web/src/shared/api/generated/GatewayApiKeySecretResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GatewayApiKeyResponse } from "./GatewayApiKeyResponse";

export type GatewayApiKeySecretResponse = { config_revision: number, items: Array<GatewayApiKeyResponse>, token: string, };